pyproject.toml
[project]
name = "congress-prioritization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google>=3.0.0",
    "google-generativeai>=0.8.6",
    "langgraph>=1.0.5",
    "litellm>=1.80.10",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
